+++ js/comments.js
@@ -1,0 +1,32 @@
+import {getRandom} from './utils.js';
+export const COUNT_OBJECTS = 25;
+
+const NAMES = [
+  'Рейнира',
+  'Дейман',
+  'Мортимер Смит',
+  'Дейнерис Бурерожденная из дома Таргариенов, законная наследница Железного Трона',
+  'Эола Лоуренс'
+];
+
+const MESSAGES = ['Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+
+export const commentsArray = (count) => {
+  const array = [];
+  for (let i = 0; i < count; i++){
+    array.push({
+      id: i,
+      avatar: `img/avatar-${getRandom(1, 6)}.svg`,
+      message: MESSAGES[getRandom(0, MESSAGES.length - 1)],
+      name: NAMES[getRandom(0, NAMES.length - 1)]
+    });
+  }
+  return array;
+};
